Move copy.determineManifestConversion to a separate file
It is fairly stand-alone, and this shortens copy.go again.

Does not change behavior (in fact, does not change the function at all).
